fix refs #2632: Keep Lark relay services behind platform composition
Move Lark-specific tail delivery, response classification, and approval notifications out of the Nyx relay channel so non-Lark relay hosts do not require Lark platform services.
