+++ agents/platforms/Aevatar.GAgents.Platform.Lark/DependencyInjection/LarkPlatformServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Aevatar.AI.ToolProviders.NyxId;
 using Aevatar.GAgents.Channel.Abstractions;
+using Aevatar.GAgents.Channel.Runtime;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
@@ -37,6 +38,12 @@ public static class LarkPlatformServiceCollectionExtensions
         services.TryAddSingleton<ILarkOutboundDispatcher, LarkOutboundDispatcher>();
         services.TryAddSingleton<LarkChannelNativeDeliveryTargetAdapter>();
         services.TryAddSingleton<LarkChannelNativeMessageSender>();
+        services.TryAddSingleton<LarkChannelRelayTailTextSender>();
+        services.TryAddSingleton<LarkRelayProxyResponseClassifier>();
+        services.Replace(ServiceDescriptor.Singleton<IChannelRelayTailTextSender>(
+            sp => sp.GetRequiredService<LarkChannelRelayTailTextSender>()));
+        services.Replace(ServiceDescriptor.Singleton<IChannelRelayProxyResponseClassifier>(
+            sp => sp.GetRequiredService<LarkRelayProxyResponseClassifier>()));
         services.TryAddEnumerable(ServiceDescriptor.Singleton<IMessageComposer, LarkMessageComposer>(
             sp => sp.GetRequiredService<LarkMessageComposer>()));
         services.TryAddEnumerable(ServiceDescriptor.Singleton<IChannelNativeMessageProducer, LarkChannelNativeMessageProducer>(
